requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
Markdown==3.7
django-filter==24.3
